Most of #13, settings for record limits
This is not yet fully complete because I can't successfully inject
`ldp.defaultShow` into the react-final-form `initialValues`. I _think_
what's going on is that because we first invoke the `<Form>` before we
know this value, it remembers the first value we've given it. This
will need some thought.

I have left debugging `console.log` lines in for now, but they will
obviously go before this is merged.