Reinstate counter as a way to force re-render
I thought we didn't need this because a race-condition meant that we
_usually_ don't: when the config-fetching `useEffect` completes before
the tables-fetching one, the `setState` in the latter forces the
re-render. But when that one finishes first, the config-fetcher just
assigns elements into the LDP context, which Reach doesn't notice --
so there is no re-render, and the app just sits there blankly.

It would be nice if React's `useContext` returned a `[value,
setFunction]` pair like `useState` does; but since it does not, we
have to find a workaround. This one seemed the clearest.