Use full width of page for results list (#43)
This turned out to be _way_ more complicated than I expected, due to the strange interactions between stripes-components' `<Paneset>` and `<Pane>` with `<div>`s  of various styles and the `<MultiColumnList>`. The good news is that the code is now much more sanely structured, though it still contains far more magic than I would like.

I also added a `<Playground>` component in with the other top-level routes. This was really helpful in figuring out how to get the display I wanted, but since it's of no interest to end-users, the link to it is now disabled (as is the link to `<LogsRoute>`.)

Fixes #28.