Replace en_US.json with an empty JSON object
It turns out that, while Stripes won't run properly under the `en_US`
locale without an `en_US.json` translation file, that file (so long as
it's present) doesn't have to contain any actual translation
keys. While it won't fall back to using `en.json` in the absence of a
file, it seems that it _will_ fall back to that default file whenever
asked to furnish a specific translation that `en_US.json` lacks. So a
version of the file that just lacks everything is fine.

"And there was much rejoicing."