Add *proof of concept* French translation
This is a horrible, horrible translation, done using Google Translate,
and should definitely not be used in real life. It's a placeholder so
I can see the results of the internationalization work in action, and
should be replaced by a real translation, done by someone who speaks
French, before this is used for real in France.