Add tests using cypress
Code coverage is currently at 87.31% statements

The big gap is that we don't exercise the error-handling code, because we have no way to provoke an error. Future work items (to be filed as their own issues) will include providing mocks for the server (so we're not reliant on a real back-end), and setting up special conditions (e.g. magic search terms) that provoke the mocks to error out.

But for now, we're good to go.

Fixes #32.