Allow for non-display of password and GitHub token
With MODLDP-25 complete, these values can still be set by PUT to
`/ldp/config/KEY`, but are omitted in responses to a GET of those
resources.

This issue changes the display to note in captions that the values are
write-only, and to put a "[Hidden]" placeholder into the relevant
editable fields.

Fixes UILDP-42.