Fake read-only status for GitHub token (#135)
This allows us to avoid displaying the GitHub token that mod-ldp has
stored and returned to us, while still allowing it to be edited. To
the user, the result is as with the DB Config page's password, but
unlike that password we do need the server to actually give the client
the GitHub token, otherwise we can't use saved searches.