Upgrade from Stripes v7 to Stripes v8 (#182)
(Though this should still work with v6 and v7).

For subtle reasons, to with `rxjs` versions, to get this to work it
was necessary to downgrade Cypress from `^7.4.0` to `^6.4.0`. This
will do for now, but as Cypress has leapt ahead five more major
versions to v12.5.1 at the time of writing, we will clearly need to
revisit this.

Fixes UILDP-79.