Correctly read configuration from mod-settings. (#259)
This includes the set of disabled tables and the default result sizes.

I messed this up when we switched from using mod-configuration to
mod-settings. We were fetching from the correct path, but incorrectly
still looking at the response data's `configs`. For mod-settings, it's
`items`.

(While I was here, I patched around a picky console error, where
`<ConfirmationModal>`'s propTypes incorrectly claim it needs a string
for the `heading` prop, when in fact it's just fine with a rendered
component.)

Fixes UILDP-144.