[UILDP-20] Export results for Excel
To do this, I made a local copy (for now) of the existing stripes-components function `exportToCSV`, and refactored it into a format-agnostic function `exportToSpreadsheet`, with the new `exportToCSV` becoming a simple front-end to that function which injects the CSV-specific functionality. I then added an analogous `exportToExcel`, so that the bulk of the code is shared between these two facilities.

Once this has been proven to work as intended, I will offer the refactored CSV export, and the accompanying Excel export, back as a pull request for stripes-components. I have no idea how that will go. It might just be gratefully accepted — in which case, great, I can remove the local copy of this code; or it might become a months-long death-march to get it through review — in which case I'll just shrug and leave it as it, with the local code. Either way is fine so far as the present issue is concerned.

I used the `xlsx` library to do the Excel-specific part of this. See the extensive discussion in UILDP-20 for details of this and of the candidate implementation options.